refactor(theta,tuple): remove internal metadata re-export
